fix: resolve all 24 x64 warnings and add LTCG to ripent

+++ src/zhlt-vluzacn/common/cmdlib.cpp
@@ -210,7 +210,7 @@ void ExtractFile(const char* const path, char* dest)
 	int extension_pos, directory_pos;
 	getFilePositions(path,&extension_pos,&directory_pos);
 
-	int length = strlen(path);
+	int length = (int)strlen(path);
 
 #ifdef ZHLT_FILE_FUNCTIONS_FIX
 	length -= directory_pos + 1;
@@ -227,7 +227,7 @@ void ExtractFileBase(const char* const path, char* dest)
 {
 	int extension_pos, directory_pos;
 	getFilePositions(path,&extension_pos,&directory_pos);
-	int length = extension_pos == -1 ? strlen(path) : extension_pos;
+	int length = extension_pos == -1 ? (int)strlen(path) : extension_pos;
 
 #ifdef ZHLT_FILE_FUNCTIONS_FIX
 	length -= directory_pos + 1;
@@ -246,7 +246,7 @@ void ExtractFileExtension(const char* const path, char* dest)
 	getFilePositions(path,&extension_pos,&directory_pos);
 	if(extension_pos != -1)
 	{
-		int length = strlen(path) - extension_pos;
+		int length = (int)strlen(path) - extension_pos;
 	    memcpy(dest,path+extension_pos,length); //include extension '.'
 	    dest[length] = 0;
 	}
