+++ src/zhlt-vluzacn/hlrad/transfers.cpp
@@ -23,7 +23,7 @@ void            writetransfers(const char* const transferfile, const long total_
     file = fopen(transferfile, "w+b");
     if (file != NULL)
     {
-        unsigned        amtwritten;
+        size_t          amtwritten;
         patch_t*        patch;
 
         Log("Writing transfers file [%s]\n", transferfile);
@@ -119,7 +119,7 @@ bool            readtransfers(const char* const transferfile, const long numpatc
     file = fopen(transferfile, "rb");
     if (file != NULL)
     {
-        unsigned        amtread;
+        size_t          amtread;
         patch_t*        patch;
 
         Log("Reading transfers file [%s]\n", transferfile);
@@ -164,7 +164,7 @@ bool            readtransfers(const char* const transferfile, const long numpatc
 				if(g_rgb_transfers)
 				{
 	#ifdef HLRAD_TRANSFERDATA_COMPRESS
-                    patch->tRGBData = (rgb_transfer_data_t*)AllocBlock(patch->iData * vector_size[g_rgbtransfer_compress_type] + unused_size);
+					patch->tRGBData = (rgb_transfer_data_t*)AllocBlock((unsigned long)(patch->iData * vector_size[g_rgbtransfer_compress_type] + unused_size));
 	#else
                     patch->tRGBData = (rgb_transfer_data_t*)AllocBlock(patch->iData * sizeof(rgb_transfer_data_t *)); //wrong? --vluzacn
 	#endif
@@ -178,7 +178,7 @@ bool            readtransfers(const char* const transferfile, const long numpatc
 				else
 				{
 	#ifdef HLRAD_TRANSFERDATA_COMPRESS
-                    patch->tData = (transfer_data_t*)AllocBlock(patch->iData * float_size[g_transfer_compress_type] + unused_size);
+					patch->tData = (transfer_data_t*)AllocBlock((unsigned long)(patch->iData * float_size[g_transfer_compress_type] + unused_size));
 	#else
                     patch->tData = (transfer_data_t*)AllocBlock(patch->iData * sizeof(transfer_data_t *));
 	#endif
